Module for synchronizing with a simulator
This patch adds an interface for pacing the execution of QEMU to match an
external simulation clock. Its aim is to permit QEMU to be used as a module
within a larger simulation system. The updates a prior patch by changing
the synchronization channel from a UNIX domain socket to a shared memory
segment with access control via semaphores.

Signed-off-by: James J. Nutaro <nutarojj@ornl.gov>